Move Circuit, Service and Builders to own files
This commit separates the module based on circuit and service
as opposed to 'builders'

Moves Circuit into  store::circuit::{Circuit, CircuitBuilder} and
Service store::service::{Service, ServuceBuilder}

The structs are still exposed only in store:: via pub. This change
also requires getters to be implemented on the structs to be used
throughout the store implementations.
