+++ libsplinter/src/admin/store/builders.rs
@@ -12,208 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Structs for building the native representations of circuits, proposals, services and nodes
+//! Structs for building services
 
 use crate::admin::messages::{is_valid_circuit_id, is_valid_service_id};
 
 use super::error::BuilderError;
 use super::{
-    AuthorizationType, Circuit, CircuitNode, CircuitProposal, DurabilityType, PersistenceType,
-    ProposalType, ProposedCircuit, ProposedNode, ProposedService, RouteType, Service, VoteRecord,
+    AuthorizationType, CircuitNode, CircuitProposal, DurabilityType, PersistenceType, ProposalType,
+    ProposedCircuit, ProposedNode, ProposedService, RouteType, VoteRecord,
 };
-
-/// Builder to be used to build a `Circuit`
-#[derive(Default, Clone)]
-pub struct CircuitBuilder {
-    circuit_id: Option<String>,
-    roster: Option<Vec<Service>>,
-    members: Option<Vec<String>>,
-    auth: Option<AuthorizationType>,
-    persistence: Option<PersistenceType>,
-    durability: Option<DurabilityType>,
-    routes: Option<RouteType>,
-    circuit_management_type: Option<String>,
-}
-
-impl CircuitBuilder {
-    /// Creates a new circuit builder
-    pub fn new() -> Self {
-        CircuitBuilder::default()
-    }
-
-    /// Returns the circuit ID in the builder
-    pub fn circuit_id(&self) -> Option<String> {
-        self.circuit_id.clone()
-    }
-
-    /// Returns the list of services in the builder
-    pub fn roster(&self) -> Option<Vec<Service>> {
-        self.roster.clone()
-    }
-
-    /// Returns the list of node IDs in the builder
-    pub fn members(&self) -> Option<Vec<String>> {
-        self.members.clone()
-    }
-
-    /// Returns the authorization type in the builder
-    pub fn auth(&self) -> Option<AuthorizationType> {
-        self.auth.clone()
-    }
-
-    /// Returns the persistence type in the builder
-    pub fn persistence(&self) -> Option<PersistenceType> {
-        self.persistence.clone()
-    }
-
-    /// Returns the durability type in the builder
-    pub fn durability(&self) -> Option<DurabilityType> {
-        self.durability.clone()
-    }
-
-    /// Returns the routing type in the builder
-    pub fn routes(&self) -> Option<RouteType> {
-        self.routes.clone()
-    }
-
-    /// Returns the circuit management type in the builder
-    pub fn circuit_management_type(&self) -> Option<String> {
-        self.circuit_management_type.clone()
-    }
-
-    /// Sets the circuit ID
-    ///
-    /// # Arguments
-    ///
-    ///  * `circuit_id` - The unique ID of the circuit
-    pub fn with_circuit_id(mut self, circuit_id: &str) -> CircuitBuilder {
-        self.circuit_id = Some(circuit_id.into());
-        self
-    }
-
-    /// Sets the list of services that are included in the circuit
-    ///
-    /// # Arguments
-    ///
-    ///  * `services` - List of services
-    pub fn with_roster(mut self, services: &[Service]) -> CircuitBuilder {
-        self.roster = Some(services.into());
-        self
-    }
-
-    /// Sets the list of node IDs for the members in the circuit
-    ///
-    /// # Arguments
-    ///
-    ///  * `members` - List of node IDs
-    pub fn with_members(mut self, members: &[String]) -> CircuitBuilder {
-        self.members = Some(members.into());
-        self
-    }
-
-    /// Sets the authorization type
-    ///
-    /// # Arguments
-    ///
-    ///  * `auth` - The authorization type for the circuit
-    pub fn with_auth(mut self, auth: &AuthorizationType) -> CircuitBuilder {
-        self.auth = Some(auth.clone());
-        self
-    }
-
-    /// Sets the persistence type
-    ///
-    /// # Arguments
-    ///
-    ///  * `persistence` - The persistence type for the circuit
-    pub fn with_persistence(mut self, persistence: &PersistenceType) -> CircuitBuilder {
-        self.persistence = Some(persistence.clone());
-        self
-    }
-
-    /// Sets the durabilitye type
-    ///
-    /// # Arguments
-    ///
-    ///  * `durability` - The durability type for the circuit
-    pub fn with_durability(mut self, durability: &DurabilityType) -> CircuitBuilder {
-        self.durability = Some(durability.clone());
-        self
-    }
-
-    /// Sets the routing type
-    ///
-    /// # Arguments
-    ///
-    ///  * `route_type` - The routing type for the circuit
-    pub fn with_routes(mut self, route_type: &RouteType) -> CircuitBuilder {
-        self.routes = Some(route_type.clone());
-        self
-    }
-
-    /// Sets the circuit management type
-    ///
-    /// # Arguments
-    ///
-    ///  * `circuit_management_type` - The circuit management type for a circuit
-    pub fn with_circuit_management_type(mut self, circuit_management_type: &str) -> CircuitBuilder {
-        self.circuit_management_type = Some(circuit_management_type.into());
-        self
-    }
-
-    /// Builds a `Circuit`
-    ///
-    /// Returns an error if the circuit ID, roster, members or circuit management
-    /// type are not set.
-    pub fn build(self) -> Result<Circuit, BuilderError> {
-        let circuit_id = match self.circuit_id {
-            Some(circuit_id) if is_valid_circuit_id(&circuit_id) => circuit_id,
-            Some(circuit_id) => {
-                return Err(BuilderError::InvalidField(format!(
-                    "circuit_id is invalid ({}): must be an 11 character string composed of two, \
-                     5 character base62 strings joined with a '-' (example: abcDE-F0123)",
-                    circuit_id,
-                )))
-            }
-            None => return Err(BuilderError::MissingField("circuit_id".to_string())),
-        };
-
-        let roster = self
-            .roster
-            .ok_or_else(|| BuilderError::MissingField("roster".to_string()))?;
-
-        let members = self
-            .members
-            .ok_or_else(|| BuilderError::MissingField("members".to_string()))?;
-
-        let auth = self.auth.unwrap_or_else(|| AuthorizationType::Trust);
-
-        let persistence = self.persistence.unwrap_or_else(PersistenceType::default);
-
-        let durability = self
-            .durability
-            .unwrap_or_else(|| DurabilityType::NoDurability);
-
-        let routes = self.routes.unwrap_or_else(RouteType::default);
-
-        let circuit_management_type = self
-            .circuit_management_type
-            .ok_or_else(|| BuilderError::MissingField("circuit_management_type".to_string()))?;
-
-        let create_circuit_message = Circuit {
-            id: circuit_id,
-            roster,
-            members,
-            auth,
-            persistence,
-            durability,
-            routes,
-            circuit_management_type,
-        };
-
-        Ok(create_circuit_message)
-    }
-}
 
 /// Builder to be used to build a `ProposedCircuit` which will be included in a `CircuitProposal`
 #[derive(Default, Clone)]
@@ -451,25 +258,6 @@ impl ProposedCircuitBuilder {
         };
 
         Ok(create_circuit_message)
-    }
-}
-
-impl From<ProposedCircuit> for Circuit {
-    fn from(circuit: ProposedCircuit) -> Self {
-        Circuit {
-            id: circuit.circuit_id,
-            roster: circuit.roster.into_iter().map(Service::from).collect(),
-            members: circuit
-                .members
-                .iter()
-                .map(|node| node.node_id.to_string())
-                .collect(),
-            auth: circuit.authorization_type,
-            persistence: circuit.persistence,
-            durability: circuit.durability,
-            routes: circuit.routes,
-            circuit_management_type: circuit.circuit_management_type,
-        }
     }
 }
 
@@ -783,117 +571,6 @@ impl From<ProposedNode> for CircuitNode {
     }
 }
 
-/// Builder for creating a `Service`
-#[derive(Default, Clone)]
-pub struct ServiceBuilder {
-    service_id: Option<String>,
-    service_type: Option<String>,
-    allowed_nodes: Option<Vec<String>>,
-    arguments: Option<Vec<(String, String)>>,
-}
-
-impl ServiceBuilder {
-    /// Creates a new `ServiceBuilder`
-    pub fn new() -> Self {
-        ServiceBuilder::default()
-    }
-
-    /// Returns the service specific service ID
-    pub fn service_id(&self) -> Option<String> {
-        self.service_id.clone()
-    }
-
-    /// Returns the service type
-    pub fn service_type(&self) -> Option<String> {
-        self.service_type.clone()
-    }
-
-    /// Returns the list of allowed nodes the service can connect to
-    pub fn allowed_nodes(&self) -> Option<Vec<String>> {
-        self.allowed_nodes.clone()
-    }
-
-    /// Returns the list of arguments for the service
-    pub fn arguments(&self) -> Option<Vec<(String, String)>> {
-        self.arguments.clone()
-    }
-
-    /// Sets the service ID
-    ///
-    /// # Arguments
-    ///
-    ///  * `service_id` - The unique service ID for service
-    pub fn with_service_id(mut self, service_id: &str) -> ServiceBuilder {
-        self.service_id = Some(service_id.into());
-        self
-    }
-
-    /// Sets the service type
-    ///
-    /// # Arguments
-    ///
-    ///  * `service_type` - The service type of the service
-    pub fn with_service_type(mut self, service_type: &str) -> ServiceBuilder {
-        self.service_type = Some(service_type.into());
-        self
-    }
-
-    /// Sets the allowed nodes
-    ///
-    /// # Arguments
-    ///
-    ///  * `allowed_nodes` - A list of node IDs the service can connect to
-    pub fn with_allowed_nodes(mut self, allowed_nodes: &[String]) -> ServiceBuilder {
-        self.allowed_nodes = Some(allowed_nodes.into());
-        self
-    }
-
-    /// Sets the service arguments
-    ///
-    /// # Arguments
-    ///
-    ///  * `arguments` - A list of key-value pairs for the arguments for the service
-    pub fn with_arguments(mut self, arguments: &[(String, String)]) -> ServiceBuilder {
-        self.arguments = Some(arguments.to_vec());
-        self
-    }
-
-    /// Builds the `Service`
-    ///
-    /// Returns an error if the service ID, service_type, or allowed nodes is not set
-    pub fn build(self) -> Result<Service, BuilderError> {
-        let service_id = match self.service_id {
-            Some(service_id) if is_valid_service_id(&service_id) => service_id,
-            Some(service_id) => {
-                return Err(BuilderError::InvalidField(format!(
-                    "service_id is invalid ({}): must be a 4 character base62 string",
-                    service_id,
-                )))
-            }
-            None => return Err(BuilderError::MissingField("service_id".to_string())),
-        };
-
-        let service_type = self
-            .service_type
-            .ok_or_else(|| BuilderError::MissingField("service_type".to_string()))?;
-
-        let allowed_nodes = self
-            .allowed_nodes
-            .ok_or_else(|| BuilderError::MissingField("allowed_nodes".to_string()))?;
-
-        let arguments = self.arguments.unwrap_or_default();
-
-        let service = Service {
-            service_id,
-            service_type,
-            allowed_nodes,
-            arguments,
-        };
-
-        Ok(service)
-    }
-}
-
 /// Builder for creating a `ProposedService`
 #[derive(Default, Clone)]
 pub struct ProposedServiceBuilder {
@@ -1002,16 +679,5 @@ impl ProposedServiceBuilder {
         };
 
         Ok(service)
-    }
-}
-
-impl From<ProposedService> for Service {
-    fn from(service: ProposedService) -> Self {
-        Service {
-            service_id: service.service_id,
-            service_type: service.service_type,
-            allowed_nodes: service.allowed_nodes,
-            arguments: service.arguments,
-        }
     }
 }
